[round-3]도메인 모델링 (#36)
* mock API 생성 - notion (#24)

* feat: 브랜드 조회 mock API 생성 (#19)

* feat: 상품 조회 mock API 생성 (#20)

* feat: 좋아요 등록/해제 mock API 생성 (#21)

* 좋아요가 된 프로젝트 목록 조회 mock API 생성 (#22)

* feat: 좋아요 등록/해제 mock API 생성

* feat: 좋아요 등록/해제 mock 조회 API 생성

* 주문 Mock API 생성 (#23)

* feat: 주문 생성 mock 조회 API 생성

* feat: 주문 목록 조회 mock API 생성

* feat: 주문 상세 조회 mock API 생성

* 단위 테스트 작업 (#31)

* 브랜드,상품,재고 단위 테스트 (#25)

* feat: 브랜드 엔티티 작성~

* feat: 상품 엔티티 작성 및 브랜드 <--> 상품 매핑

* docs: 브랜드,상품 누락되거나 추가되어진 Vo객체 추가

* docs: 재고 임베디드 타입 수정

* test: 브랜드에 프로젝트 추가 테스트

* docs: 메소드 리턴 수정

* feat: 상품,재고 엔티티 작성

* test: 상품 가격이 0보다 작을 수 없습니다.

* test: 상품,재고 단위 테스트

* 좋아요, 주문, 결제 단위 테스트 작업 (#26)

* feat: 좋아요 모델 작성

* feat: 주문, 주문아이템, 결제 모델 작성

* docs: 계정 아이디 PK 변경 AI -> 로그인 아이디

* feat: 예약어 땜에.. order -> orders로 변경

* docs: 좋아요에 삭제 시간 등록

* test: 계정 또는 상품이 존재하지 않는 경우 테스트

* test: 주문 번호 생성 테스트

* docs: 문서화 수정

* test: 주문 관련 테스트 추가

* docs: 클래스 다이어그램 수정

* test: 주문 아이템 단위 테스트

* test: 주문 히스토리 단위 테스트

* test: 결제 단위 테스트

* chore: 모든 example 제거

* 계정, 포인트 단위 테스트 작성 (#27)

* 통합 테스트 작성 (#35)

* 카탈로그 (브랜드, 상품) 통합 테스트 (#32)

* 좋아요 통합 테스트 작성 (#33)

* 주문,결제 통합 테스트 (#34)

* 결제 이후에 재고 차감 (#37)

* 주문 취소기능추가 (#38)

* refactor: 포인트,계정도 스타일 통일하기 위해 코드 수정 (#39)

* 좋아요 퍼사드 내용 수정 (#40)

+++ apps/commerce-api/src/main/java/com/loopers/application/point/PointInfo.java
@@ -1,8 +1,9 @@
 package com.loopers.application.point;
 
 import com.loopers.domain.point.PointModel;
+import java.math.BigInteger;
 
-public record PointInfo(String userId, int point) {
+public record PointInfo(String userId, BigInteger point) {
 
   public static PointInfo from(PointModel point) {
     return new PointInfo(
